pyproject.toml
[project]
name = "vllm-slurm-scheduler"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.129.0",
    "uvicorn[standard]==0.40.0",
    "httpx==0.28.1",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.0",
    "PyYAML==6.0.3",
    "SQLAlchemy==2.0.46"
]

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.15.1",
]